requirements.txt
absl-py==2.1.0
accelerate==0.27.0
annotated-types==0.6.0
astunparse==1.6.3
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blis==0.7.11
boto3==1.29.1
botocore==1.32.1
Bottleneck==1.3.7
Brotli==1.0.9
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
filelock==3.13.1
flatbuffers==24.3.25
fsspec==2024.2.0
gast==0.5.4
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
huggingface-hub==0.20.3
icrawler==0.6.7
idna==3.4
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
keras==3.3.2
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.8
networkx==3.1
numexpr==2.8.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pip==23.3.1
preshed==3.0.9
protobuf==4.25.3
psutil==5.9.8
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rich==13.7.1
s3transfer==0.7.0
sacremoses==0.0.43
sacremoses==0.0.43
sacremoses==0.0.43
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.65.0
transformers==4.37.2
typer==0.9.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.2
wheel==0.41.2
wrapt==1.16.0
